smartCollAdmSys/python-service/requirements.txt
fastapi
uvicorn
opencv-python
face_recognition
pymongo
python-multipart
python-dotenv
